Barres de progression quand on ouvre un serveur + vitesse réduite
git-svn-id: https://www.etud.insa-toulouse.fr/svn/CredigeMegaUpload/trunk@68 0ed138ed-9e3e-4e1a-b415-34bd3ada5812

+++ MegaUpload/src/org/insa/megaupload/entities/Deplacement.java
@@ -7,7 +7,7 @@ public class Deplacement {
 	private Lieu source;
 	private Lieu cible;
 	private Stack<Trajet> trajets;
-	private int avancementEtape;
+	private double avancementEtape;
 	
 	public Deplacement(Personnage perso, Lieu source, Lieu cible, Stack<Trajet> trajets) {
 		this.perso = perso;
@@ -19,14 +19,14 @@ public class Deplacement {
 	/**
 	 * @return the avancement
 	 */
-	public int getAvancementEtape() {
+	public double getAvancementEtape() {
 		return avancementEtape;
 	}
 	
 	/**
 	 * @param avancement the avancementEtape to set
 	 */
-	public void setAvancementEtape(int avancement) {
+	public void setAvancementEtape(double avancement) {
 		if (avancement >= 100) {
 			this.etapeSuivante();
 		} else {
